Move model_config to common (#71)
Creating new PR because @deadeyegoodwin is out for a bit and we want it to get it into 22.05.
Old PR
Addresses triton-inference-server/server#2759:
many of the functions in model_config.h are generically useful even on the client side and only
depend on the generated protobuf headers. Therefore, it is useful to have them more broadly
accessible via the common repo.

A few functions that have non-trivial dependence on the server API or on constants.h are kept in
the core repo, moved to model_config_utils. One definition from constants.h is moved to the common repo.

The associated changes in the model_config.h header location and the namespace are propagated
throughout the core repo.

This PR goes with triton-inference-server/common#48.