Use current time when overwriting model configuration.
This change changes the way `model_repository_manager` tracks when a model's
configuration was last updated. Instead of setting the tracked mtime to zero,
the current system time is used. This means that in order to reload the local
configuration file, a user will need to "touch" the config.pbtxt file such that
its mtime is greater than the mtime recorded when overwriting the model's
configuration values and then reload the model.

Test coverage will be added in the next commit.