Improves packaging infrastructure
Updates the Python package to use `build` to package instead of the legacy
`bdist_wheel` style. `pyproject.toml` is the single source of truth for the
package.

Additionally, this change also move stub generation to `setup.py` so that
it happens automatically when the wheels are generated. This way the stub files
don't need to be maintained in the repo and manually updated.